build: create Swift package
I created the BlockNote Swift package. The Swift package supports iOS
and macOS. I created the BlockNoteEditorView view type for UIKit and
AppKit for hosting in iOS and macOS. I created the SwiftUI
BlockNoteEditor type for iOS and macOS that wraps BlockNoteEditorView.

I created the BlockNoteSchemaHandler class that implements a URL scheme
handler for WKWebView and loads the editor's files from the Swift
package's resources.

I created a starter Vite and React application using TypeScript and
added the code to load it into the web view. I will use this web
application as the host for the BlockNote editor.

I created an example application for macOS and iOS that I can use for
testing the Swift package and to demonstrate to potential consumers how
to host BlockNote in their own applications.