feat: enable giving the web view focus
I imported the code that I wrote in another project to enable giving the
WKWebView focus without requiring user interaction. I hope to use this
to automatically give the editor focus in the future when it displays to
reduce the action that the user needs to take to start typing.